[Media Common] Share feature paramter structures between APP and driver
Share feature paramter structures between APP and driver!
