+++ media_common/agnostic/common/vp/hal/media_srcs.cmake
@@ -22,6 +22,7 @@ set(TMP_HEADERS_
     ${CMAKE_CURRENT_LIST_DIR}/vp_common.h
     ${CMAKE_CURRENT_LIST_DIR}/vp_common_hdr.h
     ${CMAKE_CURRENT_LIST_DIR}/vp_common_tools.h
+    ${CMAKE_CURRENT_LIST_DIR}/vp_common_defs.h
 )
 
 set(SOFTLET_VP_HEADERS_
